Moved common types to own file. Moved Logging to own file.
