+++ src/Microsoft.AspNet.Authentication.OAuthBearer/OAuthBearerAuthenticationMiddleware.cs
@@ -23,8 +23,6 @@ namespace Microsoft.AspNet.Authentication.OAuthBearer
     /// </summary>
     public class OAuthBearerAuthenticationMiddleware : AuthenticationMiddleware<OAuthBearerAuthenticationOptions>
     {
-        private readonly ILogger _logger;
-
         /// <summary>
         /// Bearer authentication component which is added to an HTTP pipeline. This constructor is not
         /// called by application code directly, instead it is added by calling the the IAppBuilder UseOAuthBearerAuthentication
